Move CapacityPolicies to config-privisioning
